+++ src/weather_service/DataGatherer.py
