+++ tests/data_gatherer_tests.py
@@ -1,11 +1,11 @@
 import pytest
-from src.data_gatherer.data_gatherer import DataGatherer
+from src.weather_service import DataGatherer
 
 
 def test_read_api_keys():
     test_file = "tests/test_api_key.json"
 
-    test_gatherer = DataGatherer(test_file)
+    test_gatherer = DataGatherer.DataGatherer(test_file)
 
     expected_api_keys = {
         "service": "a service",
