+++ tests/open_weather_probe_tests.py
@@ -1,0 +1,1 @@
+from src.weather_service
